src/worker/requirements.txt
redis
psycopg2-binary
python-dotenv
